+++ homepage/home.js
@@ -3,6 +3,11 @@ let navLinks=document.querySelectorAll('.navLink');
 let navBar=document.querySelector('#navBar');
 let login=document.querySelector('#login');
 login.removeAttribute('style');
+
+let dialog1=document.querySelector('#dialog1')
+let yesbtn=document.querySelector('#Yes');
+let nobtn=document.querySelector('#No');
+
 hamburger.addEventListener('click',()=>{
     hamburger.classList.toggle('active');
     navBar.classList.toggle('active');
@@ -31,11 +36,9 @@ const getUserIdFromToken = token => {
 
 if(token!==null){
   const userId = getUserIdFromToken(token);
-console.log(userId);
   fetch('https://mybrand-backend-emhu.onrender.com/users')
 .then(res=>{
   res.json().then(data=>{
-      console.log(data)
       const users=data;
       users.forEach(user=>{
 
@@ -56,20 +59,26 @@ console.log(userId);
                   localStorage.clear();
               })
               delLink.addEventListener('click',()=>{
-                  fetch(`https://mybrand-backend-emhu.onrender.com/users/${user._id}`,{
-                      method:'delete',
-                      headers:{
-                          "Content-Type":"application/json",
-                          'Authorization': `Bearer ${token}`
-                  }
-                  })
-                  .then(res=>{
-                      if(res.status===200){
-                          localStorage.clear();
-                          window.location.reload();
-                      }
-                  })
-              })
+                dialog1.showModal()
+            })
+            yesbtn.addEventListener('click',()=>{
+                fetch(`https://mybrand-backend-emhu.onrender.com/users/${user._id}`,{
+                    method:'delete',
+                    headers:{
+                        "Content-Type":"application/json",
+                        'Authorization': `Bearer ${token}`
+                }
+                })
+                .then(res=>{
+                    if(res.status===200){
+                        localStorage.clear();
+                        window.location.reload();
+                    }
+                })
+            });
+            nobtn.addEventListener('click',()=>{
+                dialog1.close()
+            })
           }
       })
   })
